Split contents of core.py:
* "create" goes to test, is a function to create images
* MegaraBaseRecipe goes to core.recipe
* The rest go to core.processing
