+++ megaradrp/core/processing.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2011-2014 Universidad Complutense de Madrid
+# Copyright 2011-2015 Universidad Complutense de Madrid
 #
 # This file is part of Megara DRP
 #
@@ -22,7 +22,6 @@ from __future__ import print_function
 from astropy.io import fits
 import numpy as np
 
-from numina.core import BaseRecipeAutoQC as MegaraBaseRecipe
 from numina.array.trace.extract import extract_simple_rss
 from numina.array.utils import wc_to_pix_1d as wcs_to_pix_1d
 
@@ -31,82 +30,8 @@ _binning = {'11': [1, 1], '21': [1, 2], '12': [2, 1], '22': [2, 2]}
 _direc = ['normal', 'mirror']
 
 
-def create(image, direction='normal', bins='11'):
-    '''Create a image with overscan for testing.'''
-
-    if direction not in _direc:
-        raise ValueError("%s must be either 'normal' or 'mirror'" % direction)
-
-    if direction == 'normal':
-        direcfun = lambda x: x
-    else:
-        direcfun = np.fliplr
-
-    if bins not in _binning:
-        raise ValueError("%s must be one if '11', '12', '21, '22'" % bins)
-
-    bng = _binning[bins]
-
-    nr = 2056 / bng[0]
-    nc = 2048 / bng[1]
-
-    nr2 = 2 * nr
-    nc2 = 2 * nc
-
-    oscan1 = 50 / bng[0]
-    oscan2 = oscan1 * 2
-
-    psc1 = 50 / bng[0]
-    psc2 = 2 * psc1
-
-    fshape = (nr2 + oscan2, nc2 + psc2)
-
-    # Row block 1
-    rb1 = slice(0, nr)
-    rb1m = slice(nr, nr + oscan1)
-    # Row block 2
-    rb2 = slice(nr + oscan2, nr2 + oscan2)
-    rb2m = slice(nr + oscan1, nr + oscan2)
-    # Col block
-    cb = slice(psc1, nc2 + psc1)
-    # Col block left
-    cbl = slice(0, psc1)
-    # Col block right
-    cbr = slice(nc2 + psc1, nc2 + psc2)
-
-    # Mode normal
-    trim1 = (rb1, cb)
-    pcol1 = (rb1, cbl)
-    ocol1 = (rb1, cbr)
-    orow1 = (rb1m, cb)
-    print(trim1, ocol1, orow1, pcol1)
-
-    trim2 = (rb2, cb)
-    pcol2 = (rb2, cbr)
-    ocol2 = (rb2, cbl)
-    orow2 = (rb2m, cb)
-    print(trim2, ocol2, orow2, pcol2)
-
-    finaldata = np.zeros(fshape, dtype='float32')
-
-    finaldata[trim1] = direcfun(np.atleast_2d(np.arange(0, nc2)))
-    finaldata[trim2] = direcfun(np.atleast_2d(np.arange(0, nc2)))
-
-    finaldata[orow1] = 3
-    finaldata[orow2] = 4
-
-    finaldata[pcol1] = 5
-    finaldata[pcol2] = 6
-
-    finaldata[ocol1] = 7
-    finaldata[ocol2] = 8
-
-    hdu = fits.PrimaryHDU(data=finaldata)
-    hdu.writeto(image, clobber=True)
-
-
 def trim_and_o(image, out='trimmed.fits', direction='normal', bins='11'):
-    '''Trim a MEGARA image with overscan.'''
+    """Trim a MEGARA image with overscan."""
 
     with fits.open(image) as hdul:
         hdu = trim_and_o_hdu(hdul[0])
@@ -114,7 +39,7 @@ def trim_and_o(image, out='trimmed.fits', direction='normal', bins='11'):
 
 
 def trim_and_o_hdu(hdu):
-    '''Trim a MEGARA HDU with overscan.'''
+    """Trim a MEGARA HDU with overscan."""
 
     # FIXME: this should come from the header
     direction = 'normal'
@@ -127,7 +52,7 @@ def trim_and_o_hdu(hdu):
 
 
 def trim_and_o_array(array, direction='normal', bins='11'):
-    '''Trim a MEGARA array with overscan.'''
+    """Trim a MEGARA array with overscan."""
 
     if direction not in _direc:
         raise ValueError("%s must be either 'normal' or 'mirror'" % direction)
@@ -160,8 +85,9 @@ def trim_and_o_array(array, direction='normal', bins='11'):
     finaldata[nr:, :] = direcfun(array[nr + oscan2:, psc1:nc2 + psc1])
     return finaldata
 
+
 def apextract(data, trace):
-    '''Extract apertures.'''
+    """Extract apertures."""
     rss = np.empty((trace.shape[0], data.shape[1]), dtype='float32')
     for idx, r in enumerate(trace):
         l = r[0]
@@ -191,7 +117,7 @@ def extract_region(data, border1, border2, pesos, xpos):
     for x, a,b in zip(xpos, border1, border2):
         fill_other(pesos[:,x], a, b)
 
-        final2d = data[region,:] * pesos[region,:]
+    final2d = data[region,:] * pesos[region,:]
 
     pesos[region,:] = 0.0
     final = final2d.sum(axis=0)
@@ -199,7 +125,7 @@ def extract_region(data, border1, border2, pesos, xpos):
 
 
 def apextract_tracemap(data, tracemap):
-    '''Extract apertures using a tracemap.'''
+    """Extract apertures using a tracemap."""
     
     # FIXME: a little hackish
     
@@ -226,7 +152,6 @@ def apextract_tracemap(data, tracemap):
     borders.append((pix_01, pix_12))
 
     rss = extract_simple_rss(data, borders)
-
 
     return rss
